Provide the default value for event attendees
When reading an event without attendees, the `attendees` has the
value of `undef`. Since this accessor is read-only, there is no
way then to populate the attendees and update the event.

This makes sure that the default value for attendees is always
an arrayref. This way, constructs like

```
@{ $entry->attendees } = @attendees;
```

are going to always work, whether we create new events or read existing events.